bonus.c split into different files
